Move endBasicGameScreenTest
